requirements.txt
aiohttp==0.20.2
aiomysql==0.0.7
alembic==0.8.5
ansible==2.0.1.0
appnope==0.1.0
astroid==1.4.4
beautifulsoup4==4.4.1
blinker==1.4
cffi==1.5.0
characteristic==14.3.0
chardet==2.3.0
colorama==0.3.6
cryptography==1.2.2
cssselect==0.9.1
decorator==4.0.6
Django==1.8.7
django-bootstrap3==7.0.1
dnspython==1.12.0
dominate==2.1.17
ecdsa==0.13
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Moment==0.5.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
gnureadline==6.3.3
html5lib==0.9999999
idna==2.0
IPy==0.83
ipykernel==4.3.1
ipyparallel==5.0.1
ipython==4.1.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
lazy-object-proxy==1.2.1
lxml==3.5.0
Mako==1.0.4
MarkupSafe==0.23
mistune==0.7.2
mysqlclient==1.3.7
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.1.0
numpy==1.10.4
pandas==0.17.1
paramiko==1.16.0
path.py==8.1.2
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.0
psutil==3.4.2
psycopg2==2.6.1
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
pycurl==7.21.5
pyflakes==1.0.0
Pygments==2.1.3
pylint==1.5.4
pymongo==3.2
PyMySQL==0.6.7
PyMySQL3==0.5
pyOpenSSL==0.15.1
python-dateutil==2.4.2
python-editor==0.5
pytz==2015.7
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.2.1
queuelib==1.4.2
redis==2.10.5
requests==2.9.1
Scrapy==1.0.4
selenium==2.53.1
service-identity==14.0.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.12
terminado==0.6
tornado==4.3
Tornado-MySQL==0.5.1
traitlets==4.1.0
tushare==0.4.7
Twisted==15.5.0
virtualenv==14.0.5
visitor==0.1.2
w3lib==1.13.0
Werkzeug==0.11.3
wrapt==1.10.6
WTForms==2.1
xlrd==0.9.4
XlsxWriter==0.8.4
zope.interface==4.1.3
